adv7511: Add more modes to HiKey whitelist
I went through a bunch more potential modes to check
if HiKey could properly drive them or not, and these
all seemed to work ok, so add them to the whitelist.

The proper fix upstream calculates if the adjusted
mode matches the pixel clock but 4.9 era kernels
are missing the required infrastructure to calculate
that here.

So for now, just extend the whitelist.

This whitelist also adds mode clock support for
linksprite touch panel:
https://www.arrow.com/en/products/96boards-display-7/linksprite-technologies-inc

Change-Id: Ia188ef8a18d0e15c1fd19a3d3f7a2ddb5677b48a
Signed-off-by: John Stultz <john.stultz@linaro.org>