drm: Revert dw_drm_dsi code to 4.4 in order to enable support for hikey-panel
Apparently the logic in the dw_drm_dsi.c code to handle
multiplexing various outputs never got upstream, so when we
synced 4.9 to match the usptream code, we effectively lost
support for the dsi panel mezzanine.

This patch reverts the needed code back to 4.4 era code (with a
few small tweaks to keep things building). With this I was
able to get the hikey panel to come up ok with HDMI unplugged.

We will need to rework this and figure out how the dsi
multiplexing should properly be done to get this upstreamed, but
for now this should get 4.9 working.

Change-Id: Ib857314888c0698aa53683fa3ca2fe8dfd09f2df
Signed-off-by: John Stultz <john.stultz@linaro.org>